% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-schema.R
\name{bind_schema}
\alias{bind_schema}
\title{Bind schema to entity}
\usage{
bind_schema(syn, entity_id, schema_id)
}
\arguments{
\item{syn}{Synapse client object}

\item{entity_id}{Synapse entity synID}

\item{schema_id}{Registered Synapse schema id}
}
\description{
Bind a registered JSON schema to a Synapse entity.
}
