% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml-json-conversion.R
\name{combine_anchors_schema}
\alias{combine_anchors_schema}
\title{Combine the anchor and yaml schema}
\usage{
combine_anchors_schema(yaml_path, anchors_path, output_path = NA)
}
\arguments{
\item{yaml_path}{Path to the yaml schema}

\item{anchors_path}{Path to the anchors file}

\item{output_path}{Path to output file (default: NA creates temp file)}
}
\value{
output_path
}
\description{
Combine the anchors and yaml schema into a single file for
dereferencing.
}
