% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{create_org}
\alias{create_org}
\title{Create new organization}
\usage{
create_org(syn, name)
}
\arguments{
\item{syn}{Synapse client object}

\item{name}{Name of the organization (must be unique across Synapse)}
}
\description{
Create a new Synapse JSON schema services organization.
}
