% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml-json-conversion.R
\name{get_anchors_from_table}
\alias{get_anchors_from_table}
\title{Convert a Synapse table to yaml anchors}
\usage{
get_anchors_from_table(syn, anchor_table, anchor_path = NA)
}
\arguments{
\item{syn}{Synapse client object}

\item{anchor_table}{synID of table to create anchor list from. Must have
columns 'alias' (the name for the anchor), 'schema' (the schema id),
'version' (the version of the schema).}

\item{anchor_path}{Output path for the yaml anchor list generated. If \code{NA}
(default), will create a temporary file.}
}
\value{
Output path for the yaml anchor list generated.
}
\description{
Convert a Synapse table to yaml anchors.
}
