% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml-json-conversion.R
\name{get_id_info}
\alias{get_id_info}
\title{Get information from id}
\usage{
get_id_info(id, info = c("id", "organization", "module", "key", "version"))
}
\arguments{
\item{id}{Full id with version in the format organization-module.key-version.}

\item{info}{One of id (unversioned id in the format
"organization-module.key"), organization (Synapse schema organization),
module (module key belongs to), key (the key), version (the version).}
}
\description{
Get information from the id.
}
