% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{get_org_acl}
\alias{get_org_acl}
\title{Organization ACL}
\usage{
get_org_acl(syn, org_id)
}
\arguments{
\item{syn}{Synapse client object}

\item{org_id}{Organziation ID number}
}
\description{
Get organization's Access Control List (ACL) by organization ID
number.
}
