% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{get_org_info}
\alias{get_org_info}
\title{Lookup organization by name}
\usage{
get_org_info(syn, name)
}
\arguments{
\item{syn}{Synapse client object}

\item{name}{Name of organization}
}
\description{
Get organization information by name.
}
