% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml-json-conversion.R
\name{get_properties_yaml}
\alias{get_properties_yaml}
\title{Get yaml anchor properties}
\usage{
get_properties_yaml(yaml_path, name = "properties", as_table = TRUE)
}
\arguments{
\item{yaml_path}{Path to the yaml file.}

\item{name}{Name of the schema to return. This is needed if there are
multiple sections of the yaml file and only a single section that has the
schema needs to be returned.}

\item{as_table}{\code{FALSE} to get the property references as a named list;
\code{TRUE} to get properties as a table with the columns Key, Version, SchemaID.}
}
\description{
Get the anchor properties from a yaml file.
}
