% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register-schema.R
\name{get_registration_status}
\alias{get_registration_status}
\title{Get status of registered schema}
\usage{
get_registration_status(syn, token)
}
\arguments{
\item{syn}{Synapse object}

\item{token}{Async token}
}
\value{
Response object containing schema
}
\description{
Get status of registered schema
}
