% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{get_schema_versions}
\alias{get_schema_versions}
\title{Get Schema Versions}
\usage{
get_schema_versions(syn, org_name, schema_name)
}
\arguments{
\item{syn}{Synapse client object}

\item{org_name}{Name of organization}

\item{schema_name}{Name of schema}
}
\description{
Get first page of information about each version of a schema.
}
