% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-schema.R
\name{get_validation_results}
\alias{get_validation_results}
\title{Get validation results}
\usage{
get_validation_results(syn, entity_id)
}
\arguments{
\item{syn}{Synapse client object}

\item{entity_id}{Synapse entity synID}
}
\value{
Synapse validation results
}
\description{
Get the annotation validation results for a Synapse entity.
}
