% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-schema.R
\name{get_validation_schema}
\alias{get_validation_schema}
\title{Get validation schema}
\usage{
get_validation_schema(syn, schema_id, readable = FALSE)
}
\arguments{
\item{syn}{Synapse client object}

\item{schema_id}{Registered Synapse schema id}

\item{readable}{Formats the schema spacing for better readability (default:
\code{FALSE})}
}
\value{
string with dereferenced JSON schema
}
\description{
Get the dereferenced validation schema from a registered Synapse
JSON schema.
}
