% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-schema.R
\name{is_valid_result}
\alias{is_valid_result}
\title{Simple check for valid annotations}
\usage{
is_valid_result(results)
}
\arguments{
\item{results}{Synapse validation results for one or more entities}
}
\value{
\code{TRUE} if all \code{results} are valid, else \code{FALSE} if one or more are
invalid
}
\description{
Check if results say the file is valid or invalid
}
