% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse-setup.R
\name{login_synapse}
\alias{login_synapse}
\title{Log into Synapse}
\usage{
login_synapse(auth_token = NA)
}
\arguments{
\item{auth_token}{Synapse Personal Access Token (default: \code{NA}). If no
\code{auth_token} given, assumes local .synapseConfig.}
}
\value{
Synapse client object
}
\description{
Loads and logs into Synapse
}
\examples{
\dontrun{
syn <- login_synapse()
syn <- login_synapse(auth_token = "my_access_token")
}
}
