% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register-schema.R
\name{register_and_report}
\alias{register_and_report}
\title{Register schema and report on its status}
\usage{
register_and_report(syn, file, dryRun = FALSE)
}
\arguments{
\item{syn}{Synapse object}

\item{file}{Path to JSON Schema file}

\item{dryRun}{Should the dryRun field be set to \code{true}? Defaults to \code{FALSE}.}
}
\value{
\code{TRUE} if schema was successfully registered; \code{FALSE} otherwise.
}
\description{
Registers a schema and then monitors the asynchronous job until it completes.
}
