% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml-json-conversion.R
\name{register_and_report_yaml}
\alias{register_and_report_yaml}
\title{Register schema from yaml}
\usage{
register_and_report_yaml(syn, yaml_path, name, dryRun = FALSE)
}
\arguments{
\item{syn}{Synapse client object}

\item{yaml_path}{Path to the yaml schema}

\item{name}{Name of the yaml schema (e.g. main-schema)}

\item{dryRun}{\code{TRUE} to check if a schema can be registered in Synapse without
registering; \code{FALSE} (default) to try to register in Synapse.}
}
\description{
Convert yaml schema to JSON schema and register in Synapse.
}
