% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse-setup.R
\name{set_synapse_endpoints}
\alias{set_synapse_endpoints}
\title{Set Synapse endpoints}
\usage{
set_synapse_endpoints(syn, staging = TRUE)
}
\arguments{
\item{syn}{Synapse client object}

\item{staging}{\code{TRUE} (default) if the endpoints should be set to staging;
\code{FALSE} if the endpoints should be set to production}
}
\value{
Synapse client object
}
\description{
Set Synapse endpoints to staging or production.
}
