% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{update_org_acl}
\alias{update_org_acl}
\title{Update organization ACL}
\usage{
update_org_acl(syn, org_id, acl)
}
\arguments{
\item{syn}{Synapse client object}

\item{org_id}{Organization ID number}

\item{acl}{JSON formmated string with new ACL}
}
\description{
Update an organization's ACL. Must have appropriate ACL
permissions to update.
}
