% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml-json-conversion.R
\name{yaml_to_json_file}
\alias{yaml_to_json_file}
\title{Convert yaml to JSON schema file}
\usage{
yaml_to_json_file(yaml_path, json_path, name = NA)
}
\arguments{
\item{yaml_path}{Path to the yaml file}

\item{json_path}{JSON output file path}

\item{name}{Name of the schema to return. This is needed if there are
multiple sections of the yaml file and only a single section that has the
schema needs to be returned.}
}
\description{
Convert a yaml schema to JSON schema and write to file.
}
