% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{list_org_schemas}
\alias{list_org_schemas}
\title{List organization schemas
TODO missing token handling for more pages}
\usage{
list_org_schemas(syn, name)
}
\description{
List organization schemas
TODO missing token handling for more pages
}
